+++ finagle-core/src/main/scala/com/twitter/finagle/tracing/DestinationTracing.scala
@@ -5,17 +5,33 @@ import com.twitter.finagle.client.Transporter
 import com.twitter.util.Future
 import java.net.InetSocketAddress
 
-/**
- * [[com.twitter.finagle.ServiceFactoryProxy]] used to trace the local addr and
- * server addr.
- */
-private[finagle] class ServerDestTracingProxy[Req, Rep](self: ServiceFactory[Req, Rep])
-    extends ServiceFactoryProxy[Req, Rep](self) {
+private[finagle] object ServerDestTracingFilter {
+  val role = Stack.Role("ServerDestinationTracing")
 
-  override def apply(conn: ClientConnection): Future[Service[Req, Rep]] =
-    self(conn).map(sf => new ServerDestTracingFilter[Req, Rep](conn).andThen(sf))
+  /**
+   * Applies the [[com.twitter.finagle.tracing.ServerDestTracingFilter]].
+   */
+  def module[Req, Rep]: Stackable[ServiceFactory[Req, Rep]] =
+    new Stack.Module1[param.Tracer, ServiceFactory[Req, Rep]] {
+      val role = ServerDestTracingFilter.role
+      val description = "Record incoming requests"
+      def make(_tracer: param.Tracer, next: ServiceFactory[Req, Rep]): ServiceFactory[Req, Rep] = {
+        if (_tracer.tracer.isNull) next
+        else {
+          new ServiceFactoryProxy[Req, Rep](next) {
+            override def apply(conn: ClientConnection): Future[Service[Req, Rep]] =
+              next(conn).map { svc =>
+                new ServerDestTracingFilter(conn).andThen(svc)
+              }
+          }
+        }
+      }
+    }
 }
 
+/**
+ * [[com.twitter.finagle.Filter]] used to trace the local and server addr.
+ */
 private final class ServerDestTracingFilter[Req, Rep](conn: ClientConnection)
     extends SimpleFilter[Req, Rep] {
 
@@ -41,46 +57,54 @@ private final class ServerDestTracingFilter[Req, Rep](conn: ClientConnection)
 
 private[finagle] object ClientDestTracingFilter {
   val role = Stack.Role("EndpointTracing")
+  val ProtocolAnnotationKey = "clnt/finagle.protocol"
 
   /**
-   * $module [[com.twitter.finagle.tracing.ClientDestTracingFilter]].
+   * Applies the [[com.twitter.finagle.tracing.ClientDestTracingFilter]].
    */
   def module[Req, Rep]: Stackable[ServiceFactory[Req, Rep]] =
-    new Stack.Module2[Transporter.EndpointAddr, param.Tracer, ServiceFactory[Req, Rep]] {
+    new Stack.Module3[
+      Transporter.EndpointAddr,
+      param.ProtocolLibrary,
+      param.Tracer,
+      ServiceFactory[Req, Rep]
+    ] {
       val role = ClientDestTracingFilter.role
       val description = "Record remote address of server"
       def make(
         _addr: Transporter.EndpointAddr,
+        _protocol: param.ProtocolLibrary,
         _tracer: param.Tracer,
         next: ServiceFactory[Req, Rep]
-      ) = {
-        val param.Tracer(tracer) = _tracer
-        if (tracer.isNull) next
-        else {
-          val Transporter.EndpointAddr(addr) = _addr
-          new ClientDestTracingFilter(addr) andThen next
-        }
+      ): ServiceFactory[Req, Rep] = {
+        if (_tracer.tracer.isNull) next
+        else new ClientDestTracingFilter(_protocol.name, _addr.addr).andThen(next)
       }
     }
 }
 
 /**
- * [[com.twitter.finagle.Filter]] for clients to record the remote address of the server.
+ * [[com.twitter.finagle.Filter]] for clients to record the remote address and protocol of the server.
  * We don't log the local addr here because it's already done in the client Dispatcher.
  */
-class ClientDestTracingFilter[Req, Rep](addr: Address) extends SimpleFilter[Req, Rep] {
+private final class ClientDestTracingFilter[Req, Rep](protocol: String, addr: Address)
+    extends SimpleFilter[Req, Rep] {
+  import ClientDestTracingFilter.ProtocolAnnotationKey
+
   def apply(request: Req, service: Service[Req, Rep]): Future[Rep] = {
     val trace = Trace()
-    val rep = service(request)
-
     if (trace.isActivelyTracing) {
+      trace.recordBinary(ProtocolAnnotationKey, protocol)
+
       addr match {
-        case Address.Inet(ia, _) =>
-          trace.recordServerAddr(ia)
+        // this filter is placed lower in the stack where the endpoint is resolved
+        case Address.Inet(sa, _) =>
+          trace.recordServerAddr(sa)
+
         case _ => // do nothing for non-ip address
       }
     }
 
-    rep
+    service(request)
   }
 }