+++ finagle-thrift/src/test/scala/com/twitter/finagle/thrift/EndToEndTest.scala
@@ -243,7 +243,7 @@ class EndToEndTest extends FunSuite with ThriftTest with BeforeAndAfter {
       // Verify the count of the annotations. Order may change.
       // These are set twice - by client and server
       assert(
-        traces.collect { case Record(_, _, Annotation.BinaryAnnotation(_, _), _) => () }.size == 9
+        traces.collect { case Record(_, _, Annotation.BinaryAnnotation(_, _), _) => () }.size == 10
       )
       assert(traces.collect { case Record(_, _, Annotation.ServerAddr(_), _) => () }.size == 2)
       // With Stack, we get an extra ClientAddr because of the